etl/requirements.txt
backoff==2.2.1
elastic-transport==8.4.0
elasticsearch==7.9.1
flake8==4.0.0
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pydantic==1.10.5
pyflakes==2.4.0
python-dotenv==0.21.0
redis==4.5.1
requests==2.28.2